+++ client/src/components/DMCampaignOverview.jsx
@@ -19,7 +19,7 @@ const DMCampaignOverview = ({ campaignData, players, deadPlayers }) => {
     const [editNameMode, setEditNameMode] = useState(false);
     const [editDescriptionMode, setEditDescriptionMode] = useState(false);
     const [name, setName] = useState(campaignData?.name);
-    const [description, setDescription] = useState(campaignData?.name);
+    const [description, setDescription] = useState(campaignData?.description);
     const [inviteEmail, setInviteEmail] = useState('');
 
     const navigate = useNavigate();