+++ client/src/components/Notes.jsx
@@ -121,13 +121,6 @@ const Notes = () => {
             )
         }
 
-        if (pending) {
-            return (
-                <Box minHeight='50vh' paddingBottom='4rem' marginTop='1rem' display='flex' alignItems='center' justifyContent='center'>
-                    <CircularProgress />
-                </Box>
-            )
-        }
         return (
             <Box height='100%' overflow='auto' display='flex' flexDirection='column' gap={2}>
                 {notesData.map(note => (
@@ -178,6 +171,7 @@ const Notes = () => {
             setDefaultCharacterFilters(characters);
             setCurrentCharacterFilters(characters);
             setFilterCharacters(characters);
+            fetchData(from, to, characters);
         }
     }, [prevCharactersPending, charactersPending, charactersData]);
 
@@ -195,7 +189,6 @@ const Notes = () => {
             setTo(currentTo);
             setFilterCharacters(currentCharacterFilters);
         }
-
     }, [showFilters]);
 
     useEffect(() => {
@@ -215,11 +208,19 @@ const Notes = () => {
     useEffect(() => {
         if (!deleteNotePending && prevDeleteNotePending) {
             if (deleteNoteSuccess) enqueueSnackbar('Note successfully deleted', { variant: 'success' });
-            if (deleteNoteError) enqueueSnackbar(addNoteError, { variant: 'error' });
+            if (deleteNoteError) enqueueSnackbar(deleteNoteError, { variant: 'error' });
         }
     }, [deleteNotePending, prevDeleteNotePending, deleteNoteSuccess, deleteNoteError]);
 
     useDebouncedPending(setPending, [notesPending, charactersPending]);
+
+    if (pending) {
+        return (
+            <Box minHeight='calc(100vh - 7rem)' display='flex' justifyContent='center' alignItems='center'>
+                <CircularProgress />
+            </Box>
+        )
+    }
 
     return (
         <>
@@ -370,7 +371,7 @@ const Notes = () => {
                         flexDirection='column'
                         gap={2}
                     >
-                        {renderNotes()}
+                        {!pending && renderNotes()}
                         {!pending && notesData.length > 0 &&
                             <Paper
                                 sx={{